ci: Generate nightly workflows for alternative runners
They were generated in the root of the repo using the following command:

    dagger call -m .github workflows generate export --path .

The new workflows are only configured to run in the context of
`dagger/dagger` repository. These runners will not be available in forks
& running the same job with the `ubuntu-latest` runner doesn't make sense.

The new workflows start with `sdk-` so that it's easier to group them
when looking at metrics. Before this, a special `OR` filter had to
created which was making additional filtering impossible.

`dockerhubAuthStep` was introduced to handle runner contexts where
`$HOME/.docker/config.json` does not get created automatically. While we
could guard against this file missing, the default behaviour that we
want is to be always authenticated, otherwise we will hit 429 errors
(which is what happened when benchmarking).

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>